+++ src/main/java/net/geforcemods/securitycraft/blockentities/SecurityCameraBlockEntity.java
@@ -48,7 +48,7 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 			}
 
 			oCameraRotation = cameraRotation;
-			
+
 			if (addToRotation && cameraRotation <= 1.55F)
 				cameraRotation += rotationSpeedOption.get();
 			else